This adds globals to the line package to allow disabling of individual tag parsing formats and unit tests for the line package. I don't love using package globals for this, but it does maintain backwards compatibility with the package, and is cleaner than adding four more parameters to the lineToEvents function. Open to other suggestions. Fixes #324 

Before:
```
$ go test -bench=BenchmarkLine.
goos: darwin
goarch: amd64
pkg: github.com/prometheus/statsd_exporter
BenchmarkLineToEventsMixed1-12            142213              9420 ns/op            4778 B/op         91 allocs/op
BenchmarkLineToEventsMixed5-12             28290             40648 ns/op           23892 B/op        455 allocs/op
BenchmarkLineToEventsMixed50-12             2908            426202 ns/op          238924 B/op       4550 allocs/op
BenchmarkLineFormats/dogStatsd-12        1870549               623 ns/op             464 B/op          6 allocs/op
BenchmarkLineFormats/invalidDogStatsd-12                 1575892               710 ns/op             496 B/op          8 allocs/op
BenchmarkLineFormats/signalFx-12                         1847803               675 ns/op             496 B/op          8 allocs/op
BenchmarkLineFormats/invalidSignalFx-12                  1661490               708 ns/op             448 B/op         11 allocs/op
BenchmarkLineFormats/influxDb-12                         1805835               614 ns/op             464 B/op          7 allocs/op
BenchmarkLineFormats/invalidInfluxDb-12                  1348052               890 ns/op             800 B/op         13 allocs/op
BenchmarkLineFormats/statsd-12                           3096838               394 ns/op             176 B/op          6 allocs/op
BenchmarkLineFormats/invalidStatsd-12                    1619024               748 ns/op             432 B/op         10 allocs/op
PASS
ok      github.com/prometheus/statsd_exporter   20.487s
```

After: 
```
$ go test -bench=BenchmarkLine.
goos: darwin
goarch: amd64
pkg: github.com/prometheus/statsd_exporter
BenchmarkLineToEventsMixed1-12            131325              8444 ns/op            4778 B/op         91 allocs/op
BenchmarkLineToEventsMixed5-12             27319             40279 ns/op           23892 B/op        455 allocs/op
BenchmarkLineToEventsMixed50-12             2986            401224 ns/op          238925 B/op       4550 allocs/op
BenchmarkLineFormats/invalidStatsd-12            1604922               745 ns/op             432 B/op         10 allocs/op
BenchmarkLineFormats/dogStatsd-12                1986829               595 ns/op             464 B/op          6 allocs/op
BenchmarkLineFormats/invalidDogStatsd-12         1713426               718 ns/op             496 B/op          8 allocs/op
BenchmarkLineFormats/signalFx-12                 1893074               636 ns/op             496 B/op          8 allocs/op
BenchmarkLineFormats/invalidSignalFx-12          1715712               712 ns/op             448 B/op         11 allocs/op
BenchmarkLineFormats/influxDb-12                 1927936               621 ns/op             464 B/op          7 allocs/op
BenchmarkLineFormats/invalidInfluxDb-12          1328978               888 ns/op             800 B/op         13 allocs/op
BenchmarkLineFormats/statsd-12                   3057933               431 ns/op             176 B/op          6 allocs/op
PASS
ok      github.com/prometheus/statsd_exporter   19.469s
```

Diff:
```
$ benchstat old.out new.out
name                             old time/op    new time/op    delta
LineToEventsMixed1-12              8.47µs ± 0%   11.37µs ± 0%   ~     (p=1.000 n=1+1)
LineToEventsMixed5-12              40.8µs ± 0%    42.7µs ± 0%   ~     (p=1.000 n=1+1)
LineToEventsMixed50-12              461µs ± 0%     400µs ± 0%   ~     (p=1.000 n=1+1)
LineFormats/influxDb-12             601ns ± 0%     604ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/invalidInfluxDb-12      918ns ± 0%     895ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/statsd-12               397ns ± 0%     414ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/invalidStatsd-12        750ns ± 0%     738ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/dogStatsd-12            630ns ± 0%     616ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/invalidDogStatsd-12     739ns ± 0%     752ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/signalFx-12             658ns ± 0%     717ns ± 0%   ~     (p=1.000 n=1+1)
LineFormats/invalidSignalFx-12      749ns ± 0%     706ns ± 0%   ~     (p=1.000 n=1+1)

name                             old alloc/op   new alloc/op   delta
LineToEventsMixed1-12              4.78kB ± 0%    4.78kB ± 0%   ~     (all equal)
LineToEventsMixed5-12              23.9kB ± 0%    23.9kB ± 0%   ~     (all equal)
LineToEventsMixed50-12              239kB ± 0%     239kB ± 0%   ~     (p=1.000 n=1+1)
LineFormats/influxDb-12              464B ± 0%      464B ± 0%   ~     (all equal)
LineFormats/invalidInfluxDb-12       800B ± 0%      800B ± 0%   ~     (all equal)
LineFormats/statsd-12                176B ± 0%      176B ± 0%   ~     (all equal)
LineFormats/invalidStatsd-12         432B ± 0%      432B ± 0%   ~     (all equal)
LineFormats/dogStatsd-12             464B ± 0%      464B ± 0%   ~     (all equal)
LineFormats/invalidDogStatsd-12      496B ± 0%      496B ± 0%   ~     (all equal)
LineFormats/signalFx-12              496B ± 0%      496B ± 0%   ~     (all equal)
LineFormats/invalidSignalFx-12       448B ± 0%      448B ± 0%   ~     (all equal)

name                             old allocs/op  new allocs/op  delta
LineToEventsMixed1-12                91.0 ± 0%      91.0 ± 0%   ~     (all equal)
LineToEventsMixed5-12                 455 ± 0%       455 ± 0%   ~     (all equal)
LineToEventsMixed50-12              4.55k ± 0%     4.55k ± 0%   ~     (all equal)
LineFormats/influxDb-12              7.00 ± 0%      7.00 ± 0%   ~     (all equal)
LineFormats/invalidInfluxDb-12       13.0 ± 0%      13.0 ± 0%   ~     (all equal)
LineFormats/statsd-12                6.00 ± 0%      6.00 ± 0%   ~     (all equal)
LineFormats/invalidStatsd-12         10.0 ± 0%      10.0 ± 0%   ~     (all equal)
LineFormats/dogStatsd-12             6.00 ± 0%      6.00 ± 0%   ~     (all equal)
LineFormats/invalidDogStatsd-12      8.00 ± 0%      8.00 ± 0%   ~     (all equal)
LineFormats/signalFx-12              8.00 ± 0%      8.00 ± 0%   ~     (all equal)
LineFormats/invalidSignalFx-12       11.0 ± 0%      11.0 ± 0%   ~     (all equal)

```