@glightfoot 2020-08-10
Yeah I had a feeling, but was lazy :-) I've changed it to use a struct for line parsing with option functions to enable tag parsing