@glightfoot 2020-08-10
Benchstat:
```
$ benchstat old.out new.out
name                             old time/op    new time/op    delta
LineToEventsMixed1-12              8.67µs ±11%    8.50µs ± 3%     ~     (p=0.841 n=5+5)
LineToEventsMixed5-12              41.5µs ± 2%    41.6µs ± 3%     ~     (p=0.841 n=5+5)
LineToEventsMixed50-12              432µs ± 5%     422µs ± 2%     ~     (p=0.222 n=5+5)
LineFormats/invalidInfluxDb-12      966ns ± 8%    1046ns ± 3%   +8.24%  (p=0.032 n=5+5)
LineFormats/statsd-12               429ns ± 7%     427ns ± 4%     ~     (p=0.952 n=5+5)
LineFormats/invalidStatsd-12        756ns ± 3%     871ns ± 8%  +15.19%  (p=0.008 n=5+5)
LineFormats/dogStatsd-12            608ns ± 3%     763ns ± 4%  +25.39%  (p=0.008 n=5+5)
LineFormats/invalidDogStatsd-12     720ns ± 3%     745ns ± 6%     ~     (p=0.103 n=5+5)
LineFormats/signalFx-12             690ns ± 1%     682ns ± 3%     ~     (p=0.183 n=5+5)
LineFormats/invalidSignalFx-12      765ns ± 1%     810ns ±17%     ~     (p=0.151 n=5+5)
LineFormats/influxDb-12             723ns ± 5%     698ns ± 7%     ~     (p=0.310 n=5+5)

name                             old alloc/op   new alloc/op   delta
LineToEventsMixed1-12              4.78kB ± 0%    4.78kB ± 0%     ~     (all equal)
LineToEventsMixed5-12              23.9kB ± 0%    23.9kB ± 0%     ~     (all equal)
LineToEventsMixed50-12              239kB ± 0%     239kB ± 0%   -0.00%  (p=0.016 n=5+5)
LineFormats/invalidInfluxDb-12       800B ± 0%      800B ± 0%     ~     (all equal)
LineFormats/statsd-12                176B ± 0%      176B ± 0%     ~     (all equal)
LineFormats/invalidStatsd-12         432B ± 0%      432B ± 0%     ~     (all equal)
LineFormats/dogStatsd-12             464B ± 0%      464B ± 0%     ~     (all equal)
LineFormats/invalidDogStatsd-12      496B ± 0%      496B ± 0%     ~     (all equal)
LineFormats/signalFx-12              496B ± 0%      496B ± 0%     ~     (all equal)
LineFormats/invalidSignalFx-12       448B ± 0%      448B ± 0%     ~     (all equal)
LineFormats/influxDb-12              464B ± 0%      464B ± 0%     ~     (all equal)

name                             old allocs/op  new allocs/op  delta
LineToEventsMixed1-12                91.0 ± 0%      91.0 ± 0%     ~     (all equal)
LineToEventsMixed5-12                 455 ± 0%       455 ± 0%     ~     (all equal)
LineToEventsMixed50-12              4.55k ± 0%     4.55k ± 0%     ~     (all equal)
LineFormats/invalidInfluxDb-12       13.0 ± 0%      13.0 ± 0%     ~     (all equal)
LineFormats/statsd-12                6.00 ± 0%      6.00 ± 0%     ~     (all equal)
LineFormats/invalidStatsd-12         10.0 ± 0%      10.0 ± 0%     ~     (all equal)
LineFormats/dogStatsd-12             6.00 ± 0%      6.00 ± 0%     ~     (all equal)
LineFormats/invalidDogStatsd-12      8.00 ± 0%      8.00 ± 0%     ~     (all equal)
LineFormats/signalFx-12              8.00 ± 0%      8.00 ± 0%     ~     (all equal)
LineFormats/invalidSignalFx-12       11.0 ± 0%      11.0 ± 0%     ~     (all equal)
LineFormats/influxDb-12              7.00 ± 0%      7.00 ± 0%     ~     (all equal)

```